refactor: UserException을 user/service로 이동, 도메인에서 예외 의존 제거
- UserException: user/exception/ → user/service/ (TournamentException, WishException 컨벤션 통일)
- User.kt: UserException import 제거, 도메인 불변식을 check()/require()로 표현
- UserService.kt: 삭제 유저·alreadyMember 가드를 서비스 레이어에서 직접 throw
- UserTest.kt: 도메인 레이어 기대 예외를 IllegalStateException/IllegalArgumentException으로 수정,
  탈퇴 유저 시나리오는 서비스 레이어 관심사이므로 도메인 테스트에서 제거
