+++ src/main/kotlin/com/depromeet/team3/user/service/UserService.kt
@@ -2,7 +2,6 @@ package com.depromeet.team3.user.service
 
 import com.depromeet.team3.user.domain.IdentityType
 import com.depromeet.team3.user.domain.User
-import com.depromeet.team3.user.exception.UserException
 import com.depromeet.team3.user.repository.UserRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -50,6 +49,7 @@ class UserService(
         newNickname: String,
     ): User {
         val user = findById(userId)
+        user.deletedAt?.let { throw UserException.deletedUser(userId) }
         user.updateNickname(newNickname)
         return userRepository.save(user)
     }
@@ -57,6 +57,8 @@ class UserService(
     @Transactional
     fun promoteToMember(userId: UUID): User {
         val user = findById(userId)
+        user.deletedAt?.let { throw UserException.deletedUser(userId) }
+        if (user.identityType != IdentityType.GUEST) throw UserException.alreadyMember(userId)
         user.promoteToMember()
         return userRepository.save(user)
     }
