+++ src/test/kotlin/com/depromeet/team3/user/domain/UserTest.kt
@@ -1,6 +1,5 @@
 package com.depromeet.team3.user.domain
 
-import com.depromeet.team3.user.exception.UserException
 import java.util.UUID
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -22,14 +21,7 @@ class UserTest {
     fun `이미 MEMBER 인 유저를 다시 승격하면 예외가 발생한다`() {
         val user = guest()
         user.promoteToMember()
-        assertFailsWith<UserException> { user.promoteToMember() }
-    }
-
-    @Test
-    fun `탈퇴한 유저를 승격하면 예외가 발생한다`() {
-        val user = guest()
-        user.softDelete()
-        assertFailsWith<UserException> { user.promoteToMember() }
+        assertFailsWith<IllegalStateException> { user.promoteToMember() }
     }
 
     @Test
@@ -49,26 +41,19 @@ class UserTest {
     @Test
     fun `닉네임 17자는 예외가 발생한다`() {
         val user = guest()
-        assertFailsWith<UserException> { user.updateNickname("12345678901234567") }
+        assertFailsWith<IllegalArgumentException> { user.updateNickname("12345678901234567") }
     }
 
     @Test
     fun `빈 닉네임은 예외가 발생한다`() {
         val user = guest()
-        assertFailsWith<UserException> { user.updateNickname("") }
+        assertFailsWith<IllegalArgumentException> { user.updateNickname("") }
     }
 
     @Test
     fun `공백만 있는 닉네임은 예외가 발생한다`() {
         val user = guest()
-        assertFailsWith<UserException> { user.updateNickname("   ") }
-    }
-
-    @Test
-    fun `탈퇴한 유저의 닉네임 변경은 예외가 발생한다`() {
-        val user = guest()
-        user.softDelete()
-        assertFailsWith<UserException> { user.updateNickname("새닉네임") }
+        assertFailsWith<IllegalArgumentException> { user.updateNickname("   ") }
     }
 
     @Test
